make cache clearing a class mixin for easier reusability
